Leaner serverless startup and request id handling
Introduce a RequestIdFilter that exposes a UUID request id to
the logging context, using AWS's request id if available in
serverless mode.

Also decouple serverless initialization from the application
initialization for a leaner startup.  Still use dropwizard to
initialize bootstrap logging, but not a jetty server etc.

Also allow multiple environments to be specified in `env`, and
add all of their config blocks as overrides.