Simplify server startup and bootstrap process
Mainly because of the KMS module initialization, we were
going through a two phased init process - first to init KMS
and then to init the rest of the app.  Instead, have all injected
config be declared in modules that install the KMS module, and
allow for multiple installations to no-op instead of throwing
exceptions by overriding equals and hashCode.

As part of this, noticed that DbModule wasn't actually properly
configurable, so remove that configuration and add a NamedDbModule
that can allow for configPath override and named binding of
database related classes.