Introduce json module and build-resource module
Instead of configuring our mapper rules all over the place, do
so in a single ObjectMapperProvider and provide a guice module
for binding it.

Put a test logback config in a module for logging during the
build process.  Can also be future home of config for build tools
like static analysis plugins, etc.