Add metrics measurement
Use metrics-guice to allow for any guice created class to
support @Timed annotaion.  Note that in kotlin, that means
the class and the method being @Timed both need to be open,
as everything is final/closed by default.

Expose ConnectionPool metrics and health checks by wiring
up the registries.

Also expose metrics, health checks, and thread dumps when
running in lambda with a configurable auth token.  These
are already exposed on the admin port when being run locally
via dropwizard.