Modify tests to allow for running in parallel
Running in parallel within a single VM instead of forking
off a VM per test runs a little faster and makes failures
a little easier to understand, so modify test that rely
on state to not do so anymore.  Also lets us allocate more
memory to the maven process since there won't be so many
of them.