Build version, deploy notification, release notes, and javadoc jars
Change the mvn build to attach javadoc jars to all projects that
generate jar artifacts.

Change the circle build to set the maven version to a relase version
for master builds, or a branch specific SNAPSHOT for other builds,
notify #code when a build is ready to be approved for deployment, and
create a github release with release notes upon successful deployment.