Refactor GraphQL WebSocket reactive code
Make a WebSocketAdapter whose job is to route websocket
events to an rx Flowable that can be subscribed to.
Make a Subscriber that handles those events and implements
the apollo protocol.  Incoming websocket commands are
handled by the Flowable/Subscriber directly.  Commands
that initiate an asynchronous action (eg, when `start`
launches db query or `connect` launches a keepalive timer)
will route the async responses back to the websocket's
Flowable, to be further handled by the Subscriber.

This allows removal of the locks that were littering the
code since routing all the websocket interaction and
subscriber state through a single Flowable/Subscriber
ensures the needed serialization.