Port subscriptions code to coroutines
Create a Channel for each WebSocket connection, and
have the WebSocketAdapter send events to the channel.
Launch a global coroutine to consume those events.
Starting queries and keepalives kick off new coroutines
that are children of the global coroutine, and get
automatically cleaned up when the global one exits.

Enable ExperimentalCoroutinesApi for Flow features,
and trigger compiler warnings to be treated as errors.
Make requestId function inline so it can be used from
both regular and suspending functions.