Cleanup GraphQL coroutines implementation
Instead of using GlobalScope, have the WebSocketAdapter
implement CoroutineScope and use that to control the
coroutines' lifecycles.  Separate out child coroutines
into their own classes that can only access the channel
and not the websocket for safety.  Also make the injected
graphQL instance non-nullable, as now that graphql is a
separate module services that don't want graphQL will
just not install the module.  Bump version to 1.6.x.