Serialize graphql json according to spec
According to the graphql spec, the data/errors/extensions
keys should exist in the json unless they have corresponding
data attached.  Some 3rd party tools like the js-graphql
plugin for jetbrains products fail to read the graphql
schema because of this deviation.  Use a jackson mixin in
order to serialize to json properly, and add the ability for
custom mixins to be injected into the ObjectMapper provider.