Support coroutine resource endpoints
Add a jersey ModelProcessor that transforms resource
methods implemented by suspend functions into a
non-suspend dynamically created function, which calls
the original suspend function and resumes an AsyncContext
with the result/exception of the coroutine.  Results in
suspend functions being supported transparently, with
an optional @AsyncDispatcher annotation to allow users
to specify which dispatcher to run the suspend function on.

h/t https://github.com/alex-shpak/rx-jersey for inspiration