Default test db connection to autoCommit=false
Since we default our runtime db connections to autoCommit=false,
do the same in the test DAO setup.  This prevents, for example,
DAO tests that save to the db but don't explicitly commit txns
from saving data to the test db when they'd not save anything at
runtime.

Breaking change, so bump version to 1.13.x