Fix coroutine jersey bugs with generics and AOP
When a resource is wrapped in an AOP dynamic subclass,
grab annotations from the definitionMethod instead of
the handlingMethod, and work around a reflection issue
by using `call(*args, continuation)` instead of
`callSuspending(*args)`.

Also use `type` instead of `rawType` to define method
parameters so that we don't lose generic type parameter
information.