Accept connectionParams on websocket for auth/context purposes
Per https://www.apollographql.com/docs/graphql-subscriptions/authentication/,
should be able to specify auth parameters in the graphql websocket's
connection_init payload.  Add a Guice configurable callback to process
the connecitonParams, so that downstreams can hook into their auth system.

Also add an implementation that integrates with dropwizard-auth as long
as the AuthFilter is available to the Guice injector.

In addition, re-authenticate credentials on the start of any
websocket query, so that a socket opened with credentials that
become invalid by the time the query is requested gets rejected.

Finally, reject any requests with an Origin that is not allowed by the
CORS config.

Bump version as auth guice bindings and websocket auth behavior have
changed.