Port to graphql-kotlin 5 / graphql-java 17 [BREAKING CHANGE]
The graphql-kotlin GraphQLContext interface is deprecated
and replaced with a new class from graphql-java that's backed
by a Map<*, Any>.  Adjust our context elements for this change.

Also some scalars are removed and moved to an -extended-scalars
package, use that for BigDecimal support.  Might want to use
more of the provided scalars in the future?