Support new graphql-ws websocket protocol
The existing subscriptions-transport-ws protocol is deprecated.
Introduce the new protocol by mapping operations and adhering
to the stricter semantics of the new protocol.  Update embedded
GraphiQL to use the new protocol.

Use existing tests to cover most corner cases, but add tests
of the new protocol's stricter semantics in addition to testing
the basic subscription workflow.