Fix GraphiQL's schema explorer when using subscriptions
The query to get schema from the server is executed with the
documentAST set to the editor pane's query instead of an AST
from the introspection query itself, and the subscription that
the fetcher returns is set up for the UI, not for internal code.
Ensure we always use a standard HTTP fetcher for the introspection
query so that the schema explorer loads properly.

https://www.notion.so/tribedynamics/Pebbles-Auto-Fill-and-Documentation-Helper-are-not-Loading-b5fc5f98a9a14c56bc74f3a85025e658