Fix coroutine requestId propogation
When executing a query over the websocket transport, we
were not getting requestIds back because things were
executing in the CoroutineScope of the connection, which
didn't yet have the requestId set on it.  Changing to
execute in a child CoroutineScope that does have the
requestId set via MDCContext() fixes the issue.

Modify the existing query-over-websocket test to
validate the query id gets set as the requestId.